+++ src/protocol.rs
@@ -1,6 +1,7 @@
 // author: kodeholic (powered by Claude)
 
 pub mod error_code;
+pub mod floor;
 pub mod message;
 pub mod opcode;
 pub mod protocol;